requirements.txt
appdirs==1.4.4
asgiref==3.4.1
black==21.7b0
click==8.0.1
dj-database-url==0.5.0
Django==3.2.5
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
gunicorn==20.1.0
Markdown==3.3.4
mypy-extensions==0.4.3
pathspec==0.9.0
PyJWT==1.7.1
pytz==2021.1
regex==2021.7.6
sqlparse==0.4.1
tomli==1.1.0
